Major restructuring to get automatic decoding/encoding
Storing raw iCalendar data or unfolded lines is silly.  We should only
store native Python types, and serialize/deserialize as needed.  This
monster commit reworks the whole package to make this possible, adding
sub-packages for a number of data-types, properties, and components
defined in RFC 5545.  The conversion is not complete yet, and there
are a number of placeholder comments waiting to be filled in still.

Some of the test output changed because now that we serialize from
scratch, we lose information about the input that was unimportant
(e.g. GEO precision is now rounded up to 6 decimals, properties are
re-ordered inside their component, etc.).